@@ -10,14 +10,14 @@
   "version": "1.1.0",
   "main": ".webpack/main",
   "scripts": {
-    "start": "electron-forge start",
+    "start": "cross-env BUILD_ENV=dev electron-forge start",
     "package": "cross-env BUILD_ENV=oss electron-forge package",
     "package-prod": "cross-env BUILD_ENV=prod electron-forge package",
     "make": "cross-env BUILD_ENV=oss electron-forge make",
     "make-prod": "cross-env BUILD_ENV=prod electron-forge make",
     "make-beta": "cross-env BUILD_ENV=beta electron-forge make",
     "publish": "electron-forge publish",
-    "lint": "npx eslint ./src",
+    "lint": "npx eslint ./src --ext .js,.jsx",
     "test": "mocha --timeout 20000 \"./test/**/*test-*.js\""
   },
   "keywords": [
@@ -55,14 +55,19 @@
     "cross-env": "^7.0.2",
     "css-loader": "^3.6.0",
     "dotenv": "^8.2.0",
-    "electron": "^9.1.2",
-    "electron-chromedriver": "^10.0.0",
-    "eslint": "^7.13.0",
+    "electron": "^11.0.3",
+    "electron-chromedriver": "^11.0.0",
+    "eslint": "^7.14.0",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-react-hooks": "^4.0.0",
     "file-loader": "^6.1.0",
     "mocha": "^8.2.0",
+    "node-abi": "^2.19.3",
     "node-loader": "^0.6.0",
-    "spectron": "^11.0.0",
+    "spectron": "^13.0.0",
     "style-loader": "^0.23.1",
     "ttf-loader": "^1.0.2"
   },
@@ -78,7 +83,7 @@
     "glob": "^7.1.6",
     "hasha": "^5.2.0",
     "jsonwebtoken": "^8.5.1",
-    "keytar": "^7.0.0",
+    "keytar": "^7.2.0",
     "ncp": "^2.0.0",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
@@ -88,6 +93,7 @@
     "react-hot-loader": "^4.12.21",
     "react-switch": "^5.0.1",
     "react-tooltip": "^4.2.10",
+    "simplebar": "^5.3.0",
     "simplebar-react": "^2.2.1",
     "stream-buffers": "^3.0.2",
     "uuid": "^8.3.0"
