@@ -30,7 +30,7 @@ rules.push({
         loader: 'babel-loader',
         options: {
             presets: ['@babel/preset-env',
-                '@babel/react', {
+                '@babel/preset-react', {
                     'plugins': ['@babel/plugin-proposal-class-properties']
                 }]
         }
@@ -40,21 +40,24 @@ rules.push({
 module.exports = {
   // Put your normal webpack config below here
   module: {
-    rules,
-    },
-    plugins: assets.map(asset => {
-        return new CopyWebpackPlugin({
-            patterns: [
-                {
-                    from: path.resolve(__dirname, 'src', asset),
-                    to: path.resolve(__dirname, '.webpack/renderer', asset)
-                },
-                {
-                    from: path.resolve(__dirname, 'src', asset),
-                    to: path.resolve(__dirname, '.webpack/renderer/main_window', asset)
-                },
+      rules,
+  },
+  plugins: assets.map(asset => {
+      return new CopyWebpackPlugin({
+          patterns: [
+              {
+                  from: path.resolve(__dirname, 'src', asset),
+                  to: path.resolve(__dirname, '.webpack/renderer', asset)
+              },
+              {
+                  from: path.resolve(__dirname, 'src', asset),
+                  to: path.resolve(__dirname, '.webpack/renderer/main_window', asset)
+              },
 
-            ]
-        });
-    })
+          ]
+      });
+  }),
+  resolve: {
+    extensions: ['.js','.jsx']
+  }
 };
